shared_mutex moved into parallel/ directory
