+++ c/read/fread/fread_thread_context.cc
@@ -5,10 +5,10 @@
 //
 // © H2O.ai 2018
 //------------------------------------------------------------------------------
-#include "read/fread/fread_thread_context.h"
 #include "csv/reader_fread.h"      // FreadReader
 #include "encodings.h"             // check_escaped_string, decode_escaped_csv_string
 #include "py_encodings.h"          // decode_win1252
+#include "read/fread/fread_thread_context.h"
 #include "read/parallel_reader.h"  // ChunkCoordinates
 #include "utils/misc.h"            // wallclock
